src/requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
streamlit
transformers
torch
tokenizers
openai
python-dotenv
nltk
scikit-learn
matplotlib
requests
biopython
tqdm